chore: Renamed classic-components to my-v1-app-classic
