--- tests/fixtures/classic-components/index.ts
@@ -1,6 +1,0 @@
-import { convertFixtureToJson } from '@codemod-utils/tests';
-
-const inputProject = convertFixtureToJson('classic-components/input');
-const outputProject = convertFixtureToJson('classic-components/output');
-
-export { inputProject, outputProject };
