+++ tests/fixtures/my-v1-app-classic/index.ts
@@ -1,0 +1,6 @@
+import { convertFixtureToJson } from '@codemod-utils/tests';
+
+const inputProject = convertFixtureToJson('my-v1-app-classic/input');
+const outputProject = convertFixtureToJson('my-v1-app-classic/output');
+
+export { inputProject, outputProject };
