+++ tests/fixtures/my-v1-app-classic/input/app/components/navigation-menu.ts
