+++ tests/fixtures/my-v1-app-classic/input/app/components/products/product/image.hbs
