+++ tests/fixtures/my-v1-app-classic/output/app/components/navigation-menu.ts
