+++ tests/fixtures/my-v1-app-classic/output/app/components/products/product/image.ts
