+++ tests/fixtures/my-v1-app-classic/output/app/components/tracks.js
