+++ tests/steps/analyze-project/my-v1-app-classic.test.ts
@@ -1,14 +1,14 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
 import { analyzeProject } from '../../../src/steps/index.js';
-import { inputProject } from '../../fixtures/classic-components/index.js';
+import { inputProject } from '../../fixtures/my-v1-app-classic/index.js';
 import {
   codemodOptions,
   context,
   options,
-} from '../../helpers/shared-test-setups/classic-components.js';
+} from '../../helpers/shared-test-setups/my-v1-app-classic.js';
 
-test('steps | analyze-project > classic-components', function () {
+test('steps | analyze-project > my-v1-app-classic', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(analyzeProject(options), context);
